+++ src/handlers/__init__.py
@@ -1,0 +1,1 @@
+from .handler import Handler
