+++ src/handlers/handler.py
@@ -1,5 +1,5 @@
 import os, json
-from .extra import find_module, install_module
+from ..utility.pip import find_module, install_module
 from typing import Any
 
 class Handler():
