+++ src/handlers/llm/gpt4all_handler.py
@@ -1,0 +1,177 @@
+import os 
+import threading
+from typing import Callable, Any
+
+from .llm import LLMHandler
+
+
+class GPT4AllHandler(LLMHandler):
+    key = "local"
+
+    def __init__(self, settings, modelspath):
+        """This class handles downloading, generating and history managing for Local Models using GPT4All library
+        """
+        self.settings = settings
+        self.modelspath = modelspath
+        self.history = {}
+        self.model_folder = os.path.join(self.modelspath, "custom_models")
+        if not os.path.isdir(self.model_folder):
+            os.makedirs(self.model_folder)
+        self.oldhistory = {}
+        self.prompts = []
+        self.model = None
+        self.session = None
+        if not os.path.isdir(self.modelspath):
+            os.makedirs(self.modelspath)
+    
+    def get_extra_settings(self) -> list:
+        models = self.get_custom_model_list()
+        default = models[0][1] if len(models) > 0 else ""
+        return [
+            {
+                "key": "streaming",
+                "title": _("Message Streaming"),
+                "description": _("Gradually stream message output"),
+                "type": "toggle",
+                "default": True,
+            },
+            {
+                "key": "custom_model",
+                "title": _("Custom gguf model file"),
+                "description": _("Add a gguf file in the specified folder and then close and reopen the settings to update"),
+                "type": "combo",
+                "default": default,
+                "values": models,
+                "folder": self.model_folder,
+            }
+        ]
+    def get_custom_model_list(self): 
+        file_list = tuple()
+        for root, _, files in os.walk(self.model_folder):
+            for file in files: 
+                if file.endswith('.gguf'):
+                    file_name = file.rstrip('.gguf')
+                    relative_path = os.path.relpath(os.path.join(root, file), self.model_folder)
+                    file_list += ((file_name, relative_path), )
+        return file_list
+
+    def model_available(self, model:str) -> bool:
+        """ Returns if a model has already been downloaded
+        """
+        from gpt4all import GPT4All
+        try:
+            GPT4All.retrieve_model(model, model_path=self.modelspath, allow_download=False, verbose=False)
+        except Exception as e:
+            return False
+        return True
+
+    def load_model(self, model:str):
+        """Loads the local model on another thread"""
+        t = threading.Thread(target=self.load_model_async, args=(model, ))
+        t.start()
+        return True
+
+    def load_model_async(self, model: str):
+        """Loads the local model"""
+        if self.model is None:
+            print(model)
+            try:
+                from gpt4all import GPT4All
+                if model == "custom":
+                    model = self.get_setting("custom_model")
+                    models = self.get_custom_model_list()
+                    if model not in models:
+                        if len(models) > 0:
+                            model = models[0][1]
+                    self.model = GPT4All(model, model_path=self.model_folder)
+                else:
+                    self.model = GPT4All(model, model_path=self.modelspath)
+                self.session = self.model.chat_session()
+                self.session.__enter__()
+            except Exception as e:
+                print("Error loading the model: ", e)
+                return False
+            return True
+
+    def download_model(self, model:str) -> bool:
+        """Downloads GPT4All model"""
+        try:
+            from gpt4all import GPT4All
+            GPT4All.retrieve_model(model, model_path=self.modelspath, allow_download=True, verbose=False)
+        except Exception as e:
+            print(e)
+            return False
+        return True
+
+    def __convert_history_text(self, history: list) -> str:
+        """Converts the given history into the correct format for current_chat_history"""
+        result = "### Previous History"
+        for message in history:
+            result += "\n" + message["User"] + ":" + message["Message"]
+        return result
+    
+    def set_history(self, prompts, history):
+        """Manages messages history"""
+        self.history = history 
+        newchat = False
+        for message in self.oldhistory:
+            if not any(message == item["Message"] for item in self.history):
+               newchat = True
+               break
+        
+        # Create a new chat
+        system_prompt = "\n".join(prompts)
+        if len(self.oldhistory) > 1 and newchat:
+            if self.session is not None and self.model is not None:
+                self.session.__exit__(None, None, None)
+                self.session = self.model.chat_session(system_prompt)
+                self.session.__enter__()
+        self.oldhistory = list()
+        for message in self.history:
+            self.oldhistory.append(message["Message"])
+        self.prompts = prompts
+
+    def generate_text_stream(self, prompt: str, history: list[dict[str, str]] = [], system_prompt: list[str] = [], on_update: Callable[[str], Any] = lambda _: None, extra_args: list = []) -> str:
+        if self.session is None or self.model is None:
+            return "Model not yet loaded..."
+        # Temporary history management
+        if len(history) > 0:
+            system_prompt.append(self.__convert_history_text(history))
+        prompts = "\n".join(system_prompt)
+        print(prompts)
+        self.session = self.model.chat_session(prompts)
+        self.session.__enter__()
+        response = self.model.generate(prompt=prompt, top_k=1, streaming=True)
+        
+        full_message = ""
+        prev_message = ""
+        for chunk in response:
+            if chunk is not None:
+                    full_message += chunk
+                    args = (full_message.strip(), ) + tuple(extra_args)
+                    if len(full_message) - len(prev_message) > 1:
+                        on_update(*args)
+                        prev_message = full_message
+        return full_message.strip()
+
+    def generate_text(self, prompt: str, history: list[dict[str, str]] = [], system_prompt: list[str] = []) -> str:
+        # History not working for text generation
+        if self.session is None or self.model is None:
+            return "Model not yet loaded..."
+        if len(history) > 0:
+            system_prompt.append(self.__convert_history_text(history)) 
+        prompts = "\n".join(system_prompt)
+        self.session = self.model.chat_session(prompts)
+        self.session.__enter__()
+        response = self.model.generate(prompt=prompt, top_k=1)
+        self.session.__exit__(None, None, None)
+        return response
+    
+    def get_suggestions(self, request_prompt: str = "", amount: int = 1) -> list[str]:
+        # Avoid generating suggestions
+        return []
+
+    def generate_chat_name(self, request_prompt: str = "") -> str:
+        # Avoid generating chat name
+        return "Chat"
+
