+++ src/handlers/stt/__init__.py
@@ -1,0 +1,1 @@
+from .stt import *
